pyproject.toml
[project]
name = "missing-file-result-board"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=6.0.2",
    "django-filter>=25.2",
    "djangorestframework>=3.16.1",
]
